Move varied_offset into builders/EUDEMO
The function is quite specific to the EUDEMO design, and the builders
will be the only things using the function.
