+++ tests/bluemira/builders/EUDEMO/first_wall/test_varied_offset.py
@@ -25,10 +25,10 @@ import numpy as np
 import pytest
 from scipy.interpolate import interp1d
 
+from bluemira.builders.EUDEMO.first_wall.varied_offset import varied_offset
 from bluemira.geometry.error import GeometryError
 from bluemira.geometry.parameterisations import PictureFrame
 from bluemira.geometry.tools import find_clockwise_angle_2d, make_circle, make_polygon
-from bluemira.geometry.varied_offset import varied_offset
 from bluemira.geometry.wire import BluemiraWire
 
 
@@ -137,7 +137,7 @@ class TestVariedOffsetFunction:
 
         Note that this is the radial offset from the center of the
         shapes, not the offset in the direction of the normal of the
-        first shape (or "normal" offset).
+        first shape (or "normal" offset, which is what we really want).
         """
         interp_1 = TestVariedOffsetFunction._interpolation_func_closed_wire(shape_1)
         interp_2 = TestVariedOffsetFunction._interpolation_func_closed_wire(shape_2)
